fix!: remove unwraps
